Ensuring the Custom Range date filter always works.
The Custom Date Range filter currently only works once. This is
because we are relying on the filter label change event to trigger
the method that updates the URL. The Date Range filter label does not
change when selecting a custom range, so the URL is never updated.

The workaround is to add an event handler for the Date Range Apply
Filter button. If the selected dates don't match the current date
range filter, we force an app-state update with the new date range.

[ST-344]