Ensured no error is returned when selecting a country filter and refreshing the page.
An "Invalid regular expression: /*/: Nothing to repeat" error was
thrown when refreshing the URL List if a country filter was selected.
This was due to a regex match() within updateFilterMenus() attempting
to match an unescaped "*" character. I added the lodash escapeRegExp()
method to ensure special regex characters are escaped before
attempting to match.

[ST-418]