Changed URL List sort field to **modifyDate**, and sort order to **desc**.
There's a service issue with sorting by Part Number, that causes the
list items to be displayed in a different order every time a user
refreshes the page. Modified Date is a more appropriate sort order,
but that field didn't exist when we initially created this page.
Modified Date sorting does not reproduce the random sorting issue.

[ST-354] [ST-352]