Ensured that the application state and URL are updated when a user changes the search field and executes a search.
When entering a search term, searching, switching the search field,
and entering a new search, the old search parameter and value remained
in the URL and application state, even though the correct search was
performed.

I added logic to the `searchQuery` property's `set()` function that
clears the value of any search field not in the `searchQuery` map.

[ST-403]