Preventing the user from using the URL to set the application-state limit higher than the Pagination component limit.
A user can use the URL to set a limit higher than 200, which could be
a problem when exporting.

I've created a List Page view model property and a check within the
component's `init()` function that sets the app-state limit to 200 if
a user attempts to set it to a higher value.

This commit also includes a cleanup of the List Page unit tests.

[ST-449]