Moved URL Title and Description Anatomy to Popover
The URL page title column appeared to be sorting incorrectly, due to the
title anatomy items displaying in the wrong order.

We will now display the full page title in the Page Title column, and
move the anatomy names and values to a popover that can be triggered by
clicking an icon next to the page title text.

These same changes will also apply to the Description field.

[ST-618]