Ensuring Page Title Whitespace is Preserved
When a page title has multiple consecutive whitespace characters, they
are collapsed into one space when displayed in the UI. If a user copies
one of these page titles and searches by it, the search will not return
the expected results.

This change ensures that the whitespace is properly displayed in the UI.

I've also removed a block of unused CSS from url-list.less.

[ST-635]