Fixing Navigation Bar Bugs and Lint Errors
There is a bug where the side navigation does not re-appear when
navigating from the request page to the home page.

We remove the Nav Bar component when `layout.hideSubheader` is **true**.
When set back to false, the component was re-inserted without the link
values, which results in the Nav Bar not rendering.

The fix is to set the `navMenuItems` property using a `get()` function,
instead of as a `value` property. Now the `navMenuItems` are
re-evaluated whenever the Nav Bar component is inserted.

There is another bug where `div#content` had a **50px** left margin when
the Nav Bar is hidden.

PUI version 1.3.5 adds a class to remove the left margin when the Nav
Bar is hidden. That class has been added to `index.stache`, and will
take effect when `layout.hideSubheader` is **true**.